requirements.txt
Flask==1.1.1
line-bot-sdk==1.17.0
Werkzeug==0.16.0
gunicorn==20.0.4
pandas==1.1.5